adding a dom helper
